return int

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
     public double calculate(double weight, double height) {
-        return weight / (height * height);
+        return (int) (weight / (height * height));
     }
 }
